pyproject.toml
[project]
name = "bcf-mock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "google-generativeai>=0.8.6",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
